Add a separate app for podcast
