+++ home/models/__init__.py
@@ -1,4 +1,3 @@
 from home.models.index import IndexPage  # noqa
 from home.models.project import ProjectPage  # noqa
 from home.models.about import AboutPage  # noqa
-from home.models.podcast import Podcast  # noqa
