+++ home/urls.py
@@ -5,5 +5,4 @@ urlpatterns = [
     path('', views.IndexView.as_view(), name="index-page"),
     path('project/', views.ProjectView.as_view(), name="project-page"),
     path('about/', views.AboutView.as_view(), name="about-page"),
-    path('podcast/', views.PodcastList.as_view(), name="podcast"),
 ]
