+++ home/views/__init__.py
@@ -1,4 +1,3 @@
 from home.views.index import IndexView  # noqa
 from home.views.project import ProjectView  # noqa
 from home.views.about import AboutView  # noqa
-from home.views.podcast import PodcastList  # noqa
