+++ podcast/admin.py
@@ -1,0 +1,5 @@
+from django.contrib import admin
+from podcast import models
+
+# Register your models here.
+admin.site.register(models.Podcast)
