+++ podcast/views.py
@@ -1,18 +1,11 @@
 from django.views import generic
 from django.db.models import Q
 
-from home.models import Podcast
-from home import forms
+from podcast.models import Podcast
+from podcast import forms
 
 
-class BasePodcastView(generic.ListView):
-    def get_context_data(self, **kwargs):
-        context = super().get_context_data(**kwargs)
-        context["form"] = forms.SearchForm(self.request.GET)
-        return context
-
-
-class PodcastList(BasePodcastView):
+class PodcastList(generic.ListView):
     model = Podcast
     context_object_name = 'podcasts'
     paginate_by = 4
@@ -25,6 +18,11 @@ class PodcastList(BasePodcastView):
                 Q(description__icontains=query) | Q(name__icontains=query),
             ).distinct()
         return queryset
+
+    def get_context_data(self, **kwargs):
+        context = super().get_context_data(**kwargs)
+        context["form"] = forms.SearchForm(self.request.GET)
+        return context
 
     def get_template_names(self):
         if self.request.headers.get('x-requested-with') == 'XMLHttpRequest':
