+++ utils/templates/extends/main.html
@@ -20,7 +20,7 @@
       <a href="/">خانه</a>
       <a href="/project">پروژه‌ها</a>
       <a href="/mentoring">منتورینگ</a>
-      <a href="/podcast">کدنالین</a>
+      <a href="{% url 'podcast:list' %}">کدنالین</a>
       <a href="/about">درباره تورهام / ارتباط</a>
       <a href="https://torhamdev.hashnode.dev/">وبلاگ</a>
       <a onclick="en_alert()" href="javascript:void(0);">EN</a>
